move save button to own file
